Add Div() extension method
