+++ CHANGELOGS.md
@@ -4,7 +4,9 @@
 [New]`BlazorRenderTree` class
 [Update]`IHasTwoWayBinding` change to `IHasValueBound`
 [Update]`HtmlHelper` optimization
+[Update]`CreateElement` and `CreateComponent` modified the arguments
 [Remove]`RenderComponentAttribute` attribute class
+
 ## 1.5.1
 [新增]`CreateElementIf`、`CreateComponentIf` 的扩展方法
 [新增]`RenderTreeBuilderFragment` 类型
