+++ common.props
@@ -1,7 +1,7 @@
 ﻿<Project>
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-    <Version>1.6</Version>
+    <Version>2.0</Version>
     <AssemblyVersion>$(Version)</AssemblyVersion>
     <FileVersion>$(Version)</FileVersion>
     <Nullable>annotations</Nullable>
